+++ packages/client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @better-upload/client
 
+## 3.0.4
+
 ## 3.0.3
 
 ## 3.0.2